requirements.txt
pandas
sqlalchemy
psycopg2
requests
pyarrow
pyyaml
python-box